+++ src/Library/Books.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
  * Created by Anthony Perez on 3/5/17.
  */
 public final class Books {
-    private static HashMap<Integer, Book> bookHash = new HashMap<>(); // ISBN, Book
+    private static HashMap<Long, Book> bookHash = new HashMap<>(); // ISBN, Book
     private static final String BOOKSFILE = "libraryBooks.csv";
 
 
@@ -25,11 +25,11 @@ public final class Books {
         ArrayList<String> authors2 = new ArrayList<>();
         authors2.add("testAuthor3");
         LocalDate date = LocalDate.now();
-        Book newBook = new Book(1, "Harry Potter and the Deathly Hallows",authors,"publisherName", date,1000,10,5);
-        Book newBook4 = new Book(2,"Harry Potter and the Prisoner of Azkaban",authors,"publisherName", date,1000,10,5);
-        Book newBook1 = new Book(3,"Catching Fire (The Second Book of the Hunger Games)",authors2,"publisherName", date,1000,10,5);
-        Book newBook2 = new Book(4,"Harry Potter Coloring Book",authors,"publisherName", date,1000,5,10);
-        Book newBook3 = new Book(5,"The Hunger Games",authors,"RIT", date,1000,10,5);
+        Book newBook = new Book(1L, "Harry Potter and the Deathly Hallows",authors,"publisherName", date,1000,10,5);
+        Book newBook4 = new Book(2L,"Harry Potter and the Prisoner of Azkaban",authors,"publisherName", date,1000,10,5);
+        Book newBook1 = new Book(3L,"Catching Fire (The Second Book of the Hunger Games)",authors2,"publisherName", date,1000,10,5);
+        Book newBook2 = new Book(4L,"Harry Potter Coloring Book",authors,"publisherName", date,1000,5,10);
+        Book newBook3 = new Book(5L,"The Hunger Games",authors,"RIT", date,1000,10,5);
         //go through file where books for the library are stored and add each book to the "bookHash".\
         LocalDate pastDate = LocalDate.now();
         //pastDate = pastDate.minusDays(1);
@@ -337,7 +337,7 @@ public final class Books {
         System.out.println(line);
     }
 
-    public static HashMap<Integer, Book> getBookHash(){
+    public static HashMap<Long, Book> getBookHash(){
         return bookHash;
     }
